Crate serial_test

serial_test

serial_test allows for the creation of serialised Rust tests using the serial attribute e.g.

#[test]
#[serial]
fn test_serial_one() {
  // Do things
}

#[test]
#[serial(some_key)]
fn test_serial_another() {
  // Do things
}

#[test]
#[parallel]
fn test_parallel_another() {
  // Do parallel things
}

Multiple tests with the serial attribute are guaranteed to be executed in serial. Ordering of the tests is not guaranteed however. Other tests with the parallel attribute may run at the same time as each other, but not at the same time as a test with serial. Tests with neither attribute may run at any time and no guarantees are made about their timing!

For cases like doctests and integration tests where the tests are run as separate processes, we also support file_serial/file_parallel, with similar properties but based off file locking. Note that there are no guarantees about one test with serial/parallel and another with file_serial/file_parallel as they lock using different methods.

#[test]
#[file_serial]
fn test_serial_three() {
  // Do things
}

All of the attributes can also be applied at a mod level and will be automagically applied to all test functions in that block

#[cfg(test)]
#[serial]
mod serial_attr_tests {
    fn foo() {
        // Won't have `serial` applied, because not a test function
        println!("Nothing");
    }

    #[test]
    fn test_bar() {
       // Will be run serially
    }
}

Feature flags

Functions

is_locked_serially

Check if the current thread is holding a serial lock

Attribute Macros

parallel

Allows for the creation of parallel Rust tests that won’t clash with serial tests

serial

Allows for the creation of serialised Rust tests