Crate serial_test

serial_test

serial_test allows for the creation of serialised Rust tests using the serial attribute e.g.

#[test]
#[serial]
fn test_serial_one() {
  // Do things
}

#[test]
#[serial]
fn test_serial_another() {
  // Do things
}

#[test]
#[parallel]
fn test_parallel_another() {
  // Do parallel things
}

Multiple tests with the serial attribute are guaranteed to be executed in serial. Ordering of the tests is not guaranteed however. Other tests with the parallel attribute may run at the same time as each other, but not at the same time as a test with serial. Tests with neither attribute may run at any time and no guarantees are made about their timing!

For cases like doctests and integration tests where the tests are run as separate processes, we also support file_serial/file_parallel, with similar properties but based off file locking. Note that there are no guarantees about one test with serial/parallel and another with file_serial/file_parallel as they lock using different methods.

#[test]
#[file_serial]
fn test_serial_three() {
  // Do things
}

Feature flags

Re-exports

pub use parallel_code_lock::local_async_parallel_core;

pub use parallel_code_lock::local_async_parallel_core_with_return;

pub use parallel_code_lock::local_parallel_core;

pub use parallel_code_lock::local_parallel_core_with_return;

pub use serial_code_lock::local_async_serial_core;

pub use serial_code_lock::local_async_serial_core_with_return;

pub use serial_code_lock::local_serial_core;

pub use serial_code_lock::local_serial_core_with_return;

Attribute Macros

parallel

Allows for the creation of parallel Rust tests that won’t clash with serial tests

serial

Allows for the creation of serialised Rust tests